ENG-14115 rename @TestLowImpactDelete to TestLowImpactDeleteNT
