+++ src/frontend/org/voltdb/SystemProcedureCatalog.java
@@ -219,7 +219,7 @@ public class SystemProcedureCatalog {
                                                                                                            false, false, false, 0,    VoltType.INVALID,   true,  false, true,  true,      false,  false,            false,        false ));
         builder.put("@NibbleDeleteSP",          new Config("org.voltdb.sysprocs.NibbleDeleteSP",           true,  false, false, 0,    VoltType.INVALID,   false, false, true,  true,      true,   false,            true,         true ));
         builder.put("@NibbleDeleteMP",          new Config("org.voltdb.sysprocs.NibbleDeleteMP",           false, false, false, 0,    VoltType.INVALID,   false, false, true,  true,      true,   false,            true,         true ));
-        builder.put("@LowImpactDelete",         new Config("org.voltdb.sysprocs.LowImpactDelete",          false, false, false, 0,    VoltType.INVALID,   false, false, false, true,      false,  false,            false,        false ));
+        builder.put("@LowImpactDeleteNT",       new Config("org.voltdb.sysprocs.LowImpactDeleteNT",        true, false, false, 0,    VoltType.INVALID,   false, false, false, true,      false,  false,            false,        false ));
         listing = builder.build();
     }
 }
